Fix failing `Select` tests with empty options
In the `Select` component (which is used as the fallback for `Select
with search`) if no options are provided then the component will not
render. There were tests that did not supply options and then weren't
able to find the (non-existant) `Select with search` component. This has
been fixed so that these tests now supply `options`.