Handle (and add explicit tests for) (force) scheduled states
We weren't handling these states before - we probably should.
See https://github.com/alphagov/whitehall/pull/10359#discussion_r2177340445

Scheduled editions are effectively 'draft' but with an automated
publish date. We should therefore treat them as drafts, and thus
reference them in the same way we reference 'editable?' edition
types.

I've added an explicit test for both scheduled and force-scheduled
editions that mirrors the 'draft' tests above it.
I've also removed the `minor_change` override in the `update_draft`
method, as it wasn't needed (it's a throwback to before we split
out the draft creation into its own `create_draft` method). I've
also had to change how we update the body, since there is a
validation rule preventing the updating of the body for scheduled
editions - so we now bypass validation when editing.