Order metadata on embedded object using schema field_order
The order in which fields are listed in a form is defined in:
`content_block_manager/config/content_block_manager.yml`
e.g.

```
field_order:
  - title
  - email_address
  - subject
  - body
  - description
```

and applied with sorting code in
`ContenBlockManager::ContentBlock::Schema`.

In this commit we implement similar code in the
`MetadataComponent` to use the same `field_order` to sort the
rows in the metadata summary list. In this way the user will
see the fields of the embedded object in the same order on
both the form and display views.

Until we have more information from users to confirm or deny
that it's appropriate to reuse the `field_order` in this
way, we won't try to DRY up this code.