Pre-compute all Edition IDs and filter out duplicates
There are a number of HTML Attachments in the migration data which
share the same parent Edition. This means that the same Edition
gets enqueued multiple times, and this we hit "There is already
an active draft for this document" errors. Pre-computing the
edition IDs and filtering out the duplicates with `.uniq` is more
efficient and avoids this race condition.

It does mean we lose a little of the logging granularity, but in
practice we're not really losing much. See
https://github.com/alphagov/whitehall/pull/10359#discussion_r2177461845