Clean up all Yard references
Remove legacy `# @!group`/`# @!endgroup` syntax.
This was added over 12 years ago in
16495bc36306254c3f576a66b018fa3d576ed2eb.

No other instances of this weird comment syntax remain - so let's
remove it. It doesn't appear to be doing anything.

As per https://github.com/alphagov/whitehall/pull/10381/files#r2194906567,
this is related to the documentation tool [Yard](https://rubygems.org/gems/yard),
which is no longer used in Whitehall, so we're tidying up all the
references.

The reason we're doing this here is that we're about to add an
`after_validation` hook in the same area, and we're tidying things
up while we're here!