Only check validity in the 'publish' context
On further inspection, the full validation suite is only run when we're
checking if we can publish the edition. Things like missing
Contact embeds are only surfaced in that context, and if we just
call `valid?` with the default context it comes out as `true`.

We therefore need to pass the `:publish` context to the `valid?`
method when setting `revalidated_at` to get an accurate result.