Add `only_invalid_editions` scope
We'll surface this in Whitehall search in a separate PR.
We want to be able to surface ALL invalid editions for a publisher
to take a look at and manually fix up.

NB I disabled the UselessAssignment rubocop rule because I think
the var names here help with the documentation.