Implement RevalidateEdition(s)Worker
The RevalidateEditionsWorker queues up all the relevant Editions
and enqueues a RevalidateEditionWorker to process each of these.
That worker simply calls `valid?(:publish)` to check the validity
of the edition for publishing, and set the cached
`revalidation_passed` boolean attribute, meaning we can easily
surface all editions with validation problems.