Run RevalidateEditionsWorker on a cronjob
We want to revalidate all Editions about once per week. We could
batch these up in chunks on a daily basis as per the
RevalidateOldLinkCheckReportsWorker, but there's as we're not
generating an expensive link check report or anything like that,
it feels pretty low risk to just batch them all up at the weekend.

That said, the TaxonValidator makes a GET request to Publishing API
so there's a chance we could overwhelm the system by accident!
Will consider this further.