Surface validation errors in the sidebar
Now that we're offering publishers a way of seeing all their
'invalid' editions directly from the search screen, we need a
better journey for them to actually find out what the validation
issues are (without them having to edit or try-and-fail-to-publish
the edition). So we now run the `valid?(:publish)` checks on
`:show` and surface any validation issues in the sidebar.

This also surfaces on non-draft editions so publishers can see
where past editions would no longer be considered 'valid', which
is quite a useful audit trail.