Add 'revalidated_at' property to editions table
We are going to introduce a background worker that cycles through
all published editions and checks whether they're still considered
'valid'. This is to catch cases where published editions are no
longer valid, such as using legacy GovSpeak or referring to a
Contact embed that no longer exists.

Running the `valid?` checks at runtime, across all editions, is
_very_ slow - so we need to pre-compute the outputs ahead of time
if we want to be able to surface invalid editions in search.

We initially considered adding a `revalidation_passed` boolean
(true for valid editions, false for invalid) but then that limits
our ability to know how recently something was revalidated, which
hampers our ability to intelligently batch editions for revalidation.
So following the advice of <https://changelog.com/posts/you-might-as-well-timestamp-it>