DRY up the `base_path` methods
Now that we have a one-way mapping of base paths to Document Types,
this commit introduces a two-way mapping where we can take a
Document Type and slug and generate a base path, DRYing up our
models in the process.

This is likely the kind of direction we'll head in for the schema
driven content types in Whitehall over the next few months, but
may be considered something of an overcomplexity in the meantime.