Make bulk-retagging feature simpler and easier to use
On the latest use of this feature (https://govuk.zendesk.com/agent/tickets/6148849),
there was a lot of back and forth trying to get the CSV exactly
right. There was confusion over the document type vs the schema
name (not helped by the fact that we're _actually_ wanting the
schema name, despite saying we want the document type, because
document_type is the property we're looking up in Whitehall).

There was all this complex "ambiguity checking" which meant that
we needed both the slug and the document type. But that is all
rendered redundant if we're happy just taking a GOV.UK URL - which
is also much easier for the publisher (no pre-processing of the
URLs is required).