Add BasePathHelper
A fairly common need is to figure out a document type from its
GOV.UK URL. Prior to this commit, we had no way of dynamically
deducing this - we'd have to look into each model manually until
we find the right one. This commit is an attempt to surface all
of the doctype/slug mappings in one place.

It covers all of the main document types, but leaves out the
composable models (e.g. Person) for the time being.

In the next commit we'll use this new helper to improve Whitehall's
bulk-retagging feature.