Fix schedule-publishing loophole that leads to missing taxon
Whitehall doesn't allow publishers to publish documents that have
not been tagged to a topic taxonomy. The 'publish' and 'force
publish' actions don't succeed.

But a workaround has been to schedule a document for publishing.
Publishers are duly told "Your document has been saved. You need
to add topic tags before you can publish this document." in the
subsequent 'success' flash message upon successfully saving or
scheduling their document, but there are no further checks made
before the edition is published. Consequently we have a number of
live editions without a topic taxonomy. This isn't great as it
affects users ability to navigate to content outside of using the
search bar, and also limits their ability to subscribe to email
alerts about a topic.

The issue seems to stem from us calling the more forgiving
`valid?` method in the edition publisher / scheduler, where we
should really be calling `valid?(:publish)` to check that the
edition is valid in a published context. The latter has more
stringent checks such as checking that a document has been
tagged to a topic taxonomy, that all of its Contact embed codes
refer to contacts that actually exist, etc.

Also removes some straggling code that should have been removed
in #10284, the last time we were in this area.

IMPACT ON THE TESTS:

We're now doing a full `valid?(:publish)` call on the edition
schedule/publish workflows. This more comprehensive validity check
uses the TaxonValidator to ensure that the document has been
tagged to a topic taxonomy. We therefore have to mock that a
document has had its pre-requisite tagging done, in order to not
break all the tests.

I did explore a few approaches, e.g. stubbing
`Services.publishing_api.get_links` to return a links hash with
taxons hash with an array of links, but then that causes a lot
of test failures where we're expecting NO links (directly, or
indirectly). So my eventual solution, which requires minimal
changes, is to stub a successful response from TaxonValidator
except in the very few cases where its behaviour is being
directly tested (in which case we 'unstub' it).