Generate object keys more intelligently
Up until now, we worked on the assumption that all subobjects would
require a title, which we would then use to generate an object key.

For example, if we added an email object to a contact with the name
“My email address”, this would be added to the `email` object with
the key `my-email-address`. For example:

```json
…

{
  “emails”: {
      “my-email-address”: { … }
}
…
```

We decided early on to make the titles mandatory, because otherwise we
would have no way of generating a predictable identifier for an
individual object.

As we’ve moved along the road, we’ve discovered that actually, mandating
the use of a title causes more problems than it solves, so we’ve decided
to make `title` an optional field and fall back to generating a key if
a title is not present.

So, in the example above, if we were to add an email to our contact, and
there was no title provided, the key would default to `email-1`.

Additionally, if there was already an email object present, we would use
the count of the current number of email objects, and increment this
count to generate the key `email-2`

We already did have some fallback code to generate a random identifier,
but this never got hit, becuase all subobjects had mandatory titles.