Add 'label' field to 'call_charges' content
This attribute completes the current design of the "call
charges" sub-object within the phone numbers embedded object
in the "Contact" content block.

Like the `call_charges_info_url` it also has a "default"
value ("Find out about call charges") which can be over-ridden
if the user needs to customise that label.