Remove unnecessary "hasn't been published" prefix
Now that we're surfacing all validation errors in the sidebar by
default, the wording "This document has not been published. You
need to add a topic before publishing." doesn't make sense in
every context. For example, a publisher viewing their draft edition
would see that message (and think, quite rightly, that they DON'T
want the document to be published yet, and perhaps think - wrongly -
that if they do add a topic, it will somehow automatically get
published).

The fact that the document didn't publish (if the publisher tries
to do so without adding a topic) should already be pretty evident
from the error flash message at the top of the page, the lack of
'published' state in the sidebar, and the continued display of
"Invalid edition" in the sidebar.