Rename Manual Content Page to Manual Section
This is closer to the actual document type, and I think slighly less
confusing. Fundamentally though I think these are always going to be
confusing because the manual "sections" are often also index pages, and
users probably expect a content type for index pages and one for content
pages, but that's not how it works.