Support nested objects in field diffs
Before the field diffs were breaking if we made a change to any nested
objects. We get around this by “flattening” the hash to a single level
hash with the keys joined with an underscore.

This is by no means perfect, but it fixes the issue for now and gets
us ready for UR