WIP: add a LegacyPresenter override
Theory being, we can migrate existing standard edition types
inside Whitehall, to the new flexible pages format, *without*
Publishing API, Content Store and the Frontend being updated to
accommodate the change. Then, when all the folks downstream of
Whitehall are ready, switching to flexible page rendering would
be as simple as removing the `legacy_presenter` property from the
schema.

Spent an hour or two on this, it's a bit fiddly, but I'm sure it's
doable.