Change `BulkUpload` to use files not zip file
Previously, bulk uploading files meant uploading a `.zip` file with the
files you wanted to upload. This commit changes `BulkUpload` to use a
list of uploaded files instead. As a result we can now use the native
browser functionality of `input[type=file][multiple]`.

The code for bulk uploading multiple files is now simplified and the
`.zip` handling code has been removed. Although `BulkUpload` is still
currently a separate route, it is now in a place where it can eventually
be merged into the existing file attachment code and replace the
"individual" file upload journey. Maintaining this as a separate route
for now means we can get feedback and limit disruption of the existing
user experience.