Refactor `BulkUpload` form fields
Remove the duplicated mark-up that existing in the `BulkUpload`
templates and use the `Attachment` templates for rendering the
form fields instead.

Restructure existing form to use `fieldset` for individual files to be
edited and rework the heading hierarchy within each individual `fieldset`.
This will make the `BulkUpload` form more accessible.