Handle name collisions in object keys
If an object has the same title, then we add a number as appropriate to
prevent name collisions. Additionally, I’ve simplified the code so if
a title is missing we use the object_type as the key.

I’ve had to tweak the tests, so the first time an object_type is used,
a number isn’t appended, but I think this makes sense in reality.