Update `translated_value` to use the key first
In certain circumstances (for example, drop downs), we don’t necessarily
want to convert all uses of words (for example `welsh_language` or
`telephone`) to their translated value. This adds a `key` argument, so
we first look up the translation in the context of the key, then fall
back to the value on its own. This allows us to do global translations
if we want (which is relevant for, say booleans), but first check in
the context of the key.