Fix up orgs superseded by "Academy for Social Justice"
**Background**
----

JIRA: https://gov-uk.atlassian.net/browse/WHIT-1447

There are several live pages under the old “Academy for Social Justice Commissioning” organisations - but none of these pages are accessible/editable in Whitehall:

- https://www.gov.uk/government/organisations/academy-for-social-justice-commissioning
- https://www.gov.uk/government/organisations/academy-for-social-justice-commissioning/about
- https://www.gov.uk/government/organisations/academy-for-social-justice-commissioning/about/our-governance
- https://www.gov.uk/government/organisations/academy-for-social-justice-commissioning/about/membership
- https://www.gov.uk/government/organisations/academy-for-social-justice-commissioning/about/about-our-services

To make matters worse, "Academy for Social Justice Commissioning" itself replaces an earlier organisation, "Academy for Justice Commissioning", available at the following org and CorporateInformationPages:

- https://www.gov.uk/government/organisations/academy-for-justice-commissioning
- https://www.gov.uk/government/organisations/academy-for-justice-commissioning/about
- https://www.gov.uk/government/organisations/academy-for-justice-commissioning/about/membership

This appears to stem from an original [request for changing the name](https://govuk.zendesk.com/agent/tickets/3587898) of the org to Academy for Social Justice.

**Why this is an issue, and what to do about it**
----

[As discussed on Slack](https://gds.slack.com/archives/C02L13S214K/p1747228874147669), orgs should never be deleted when their name changes. They should instead be marked as 'closed'. **Their CorporateInformationPages should continue to be live and editable in Whitehall**.

We therefore need to reinstate the deleted organisations manually, and resurrect their associated CorporateInformationPages too.

**Screenshots**
----

It's important to note that we will lose the Documents section of these org pages, as republishing the pages renders the page anew and the search API calls that power that section of the page now return zero documents. This is unavoidable.

Note that the org pages currently lack the new GOV.UK branding. That's because the orgs no longer exist in Whitehall, and thus any bulk-republish jobs ran as part of the rebrand work have not applied here.

Academy for Social Justice Commissioning:

- https://www.gov.uk/search/news-and-communications?organisations[]=academy-for-social-justice-commissioning&parent=academy-for-social-justice-commissioning

|Before|After|
|------|------|
|![](https://github.com/user-attachments/assets/b6aa4722-db03-4a67-9fa3-d3e541db3c12)|![](https://github.com/user-attachments/assets/330070ad-b88b-42ba-a020-afdd7b451472)|

Academy for Justice Commissioning:

  - https://www.gov.uk/search/news-and-communications?organisations[]=academy-for-justice-commissioning&parent=academy-for-justice-commissioning
  - https://www.gov.uk/search/policy-papers-and-consultations?organisations[]=academy-for-justice-commissioning&parent=academy-for-justice-commissioning
  - https://www.gov.uk/search/transparency-and-freedom-of-information-releases?organisations[]=academy-for-justice-commissioning&parent=academy-for-justice-commissioning

|Before|After|
|------|------|
|![](https://github.com/user-attachments/assets/99bf1594-1e77-4065-a8b9-a4b46af8821f)|![](https://github.com/user-attachments/assets/0bf56457-1c91-4f4b-9d19-abb6e02374df)|

This PR also required manually re-creating the "our governance" page for Academy for Social Justice Commissioning so that it became editable:

|Before|After|
|------|------|
|![](https://github.com/user-attachments/assets/10a9cde5-adaa-475a-9454-4abc7b955594)|![](https://github.com/user-attachments/assets/90c000ac-7caa-4b6a-b96e-bc02ad3aa717)|

Currently live version of [About page for Academy for Social Justice Commissioning](https://www.gov.uk/government/organisations/academy-for-social-justice-commissioning/about), showing there is no link rendered, so the markup can be safely removed:

> ![](https://github.com/user-attachments/assets/a0eeb630-6652-48f1-811e-96f8483f1285)