Make `file` validation and errors consistent
If an invalid `file` was uploaded using `AttachmentUploader`, it wouldn't
surface the correct error to the user. It would always say that the file
was `blank`. This commit fixes this so that the error messages will now
surface.

This commit also removes validation within `build_attachment_for_file`
of `BulkUpload`. It now uses the validators which are already
implemented on the `AttachmentUploader`. This means we have consistent
validation behaviour for all files.