Move 255 URL maxlength limit into the UriValidator
There's no reason this should have been applied only to the
Nation Applicability 'alternative URL' validation. It should be
encapsulated in the UriValidator for everything to benefit from.

I've also removed the duplicate tests for the inapplicability
file. It is enough to test that the model calls the uri_validator,
and not to repeat the assertions.