Fix `500` for `WorldwideOrganisation` offices
This error was occurring because `find_worldwide_organisation` used
`Edition.find` instead of `WorldwideOrganisation.find`. This meant that
an Edition could exist with the `worldwide_organisation_id` in the param
but it might be not actually be a `WorldwideOrganisation`. This would
then cause an error when `offices` was called on the `Edition`.