Remove secondary implementation of access limiting logic
The main access limiting logic for editions is implemented in the edition rules, but the access limiting module also exposed an `accessible_to` class method which provided a secondary implementation.

The class method was only called by an instance method on the same module, and that instant method was only called when checking the visibility of attachments and call for evidence and consultation responses. It shouldn't be possible to get a different result when checking the visibility of an attachment or response from its parent edition, so we remove the secondary access limiting implementation in favour of using the one specified in the edition rules, which is also a simpler implementation.