Refactor VideoRelayServiceComponent test
This makes the test more clearly describe the behaviour of
the component. We now:

- instantiate the component with a subschema rather than a
schema. Only subschemas currently have the facility to
understand whether nested objects' fields are
"govspeak-enabled".

- demonstrate how the "subschema_id" (aka
"subschema_block_type") is used to build representations of
the field's path. i.e. in this example "telephones" is the
subschema_block_type in the `name`:

`content_block/edition[details][telephones][video_relay_service][prefix]`

We do this refactoring before extracting the textarea from
this component into the new
`GovspeakEnabledTextareaComponent`.