Show Govspeak 'Preview' in action in cucumber
For now we just demonstrate that the JS is wired up
correctly: when clicking the 'Preview' button the user sees
the govspeak preview pane, populated with the initial value
of "Generating preview...".

It would be nice to see the HTML returned by the Whitehall
preview service. The JS (`GOVUK.Modules.GovspeakEditor`) sends
a `POST` to:

`http://whitehall-admin.dev.gov.uk/government/admin/preview`

and receives an HTML fragment in response e.g.

```html
section>
  <article class="document">
    <div class="body">
      <div class="govspeak">
        <p>
          Rich <strong>content</strong>
        </p>
      </div>
    </div>
  </article>
</section>
```

I haven't figure out where this request is being swallowed
up but probably it's sufficient to show that the govspeak
editor widget is set up correctly?