Move all module 'include' statements to top of model
I was trying to figure out what the `remove_from_search_index`
method was doing in the StatisticsAnnouncement model if it doesn't
include the Searchable module. But it turns out it does - it just
wasn't declared as such at the top of the file.

This annoyed me enough to go through the codebase and move all
of the statements to the top of the model so it's really clear
what models are dependent on what modules.