Refactor away flexible blocks context
The flexible block objects previously used a "Context" object to make the flexible page model available to the blocks so that they can do things such as access the value of other associations that belong to the page. The Context object was set up in a controller "before" action for each request.

This was a poor design, because the context was essentially a big hairy global object and resulted in blocks having access to far more information than they needed to do their jobs.

We therefore have refactored the app to give more responsibility to the block factory, which receives the flexible page model instance and is now responsible for extracting any information required from the model instance and passing it to each block in the build method. This allows us to hide the page model from the blocks.

We also take advantage of Rails magical partial rendering so that we can access the block from inside its erb template. We add attribute readers to blocks so that information passed to the block initializers can be used in the erb template.