Kill off 'Announcement' parent class
This was only being used to pull in dependent modules - which,
as discussed in previous commit, we want to explicitly list per
content type anyway, so not helpful! - and also to define a
seemingly common 'base_path' definition. But on closer inspection,
'base_path' is being overridden in FatalityNotice and in Speech,
so the only model that was actually using this value defined in
the Announcement parent was NewsArticle. We may as well define it
inline.

It looks like it was once used to power some generic search
filters and the like, but they don't appear to be surfaced in
Whitehall search anymore.

NB, I've deleted the `"without_editions_of_type doesn't exclude
subclasses of the supplied classes` test because we now no longer
have any Edition descendents that inherit from anything other than
Edition itself!