Move checking supported attachment type to model
Previously, the check for if an attachment supported an attachment type
was in `AttachmentController`. This meant that unsupported attachments
could still be created if the method of creation was outside of the
controller. This was evident by several instances of tests in the suite
that would create invalid attachment types on document types with no
validation errors.

If this is an intended restriction, it seems like it would be a better
fit within the `Attachment` model. So it has been moved there instead
and now an invalid attachable will result in an error being raised.