Stop swallowing 422 responses from Publishing API
When attempting to unpublish a document, if Whitehall's validation
rules don't align with Publishing API's validation rules, we can
get into a state whereby Whitehall thinks the document has been
successfully unpublished but Publishing API rejects the change,
and the document continues to remain live.

Instead of just swallowing the error and manually sending details
of it to Sentry, we want to bubble the error up and essentially
cause Whitehall to crash ("Server error") rather than have the
document states become out of sync.

Whilst it would be good to go the extra mile and handle the
exception properly - displaying details of the validation issue
to the user - in practice, we're only seeing this in non production
environments, due to an issue that we hope to resolve separately
in #10549. So showing the "Server error" message and forcing users
to get in touch with support seems a reasonable improvement for
now.

JIRA: https://gov-uk.atlassian.net/browse/WHIT-2401