Remove `miller-columns-element` package
As part of the work to archive `miller-columns-element`, this commmit
adds the compiled JS and CSS of the `miller-columns-element` package to
`whitehall` and removes `miller-columns-element` from `package.json`.

This is possible as there are no external JS dependencies for the
`miller-columns-element` package. So it can be safely used as a
compiled script until we are able to refactor the component.