Create data migration to fix up orphaned assets
There a number of orphaned assets on Whitehall that should be either deleted or redirected due to a previous outage on 31st July with Asset Manager.

This data migration cleans up those specific attachments that have been deleted in Whitehall but still need to be deleted from Asset Manager.

There were other instances where a new draft had been created for a previously unpublished Edition. Assets had failed to be updated with a redirect url when the draft was deleted. This migration will update those assets with the correct redirect url.