Move main office editing to edition edit
To prevent an error in which editing the main office of the draft of an
already published Worldwide Organisation would cause an error due to
a missing change note.

Main office is now editable within the main edition form of Worldwide
Organisation. This means we can remove the controller which previously
handled this edit and the associated views and tests.

The link to `Set main office` now links the user to the field within the
edition edit which controls this.