Move main office editing to edition edit
To prevent an error in which editing the main office of the draft of an
already published Worldwide Organisation would cause an error due to
a missing change note.

Main office is now editable within the main edition form of Worldwide
Organisation. This means we can remove the controller which previously
handled this edit and the associated views and tests.

In addition to preventing this error occuring, moving `Main office`
selection to `Edition#edit` better reflects the relationship of the
data (`main_office_id` is set on a WorldwideOrganisation). It also
reflects the longterm goal of reusable contacts within Whitehall from
the Content Block Manager that will not be associated with a Worldwide
Organisation.

The link to `Set main office` now links the user to the field within the
edition edit which controls this.